requirements.txt
#https://github.com/makcedward/nlpaug
ffmpeg-python
numpy 
librosa
